+++ tests/testthat/test-req-perform-connection.R
@@ -64,8 +64,39 @@ test_that("curl errors become errors", {
 
   # and captures request
   cnd <- catch_cnd(req_perform_connection(req), classes = "error")
-  expect_equal(cnd$request, req)
+  expect_equal(cnd$request, req_policies(req, connection = TRUE))
 
   # But last_response() is NULL
   expect_null(last_response())
+})
+
+test_that("mocking works", {
+  req_200 <- request("https://ok")
+  req_404 <- request("https://notok")
+
+  local_mocked_responses(function(req) {
+    expect_equal(req$policies$connection, TRUE)
+    if (req$url == "https://ok") {
+      conn <- rawConnection(charToRaw("a\nb\n"))
+      response(body = StreamingBody$new(conn))
+    } else {
+      response(404)
+    }
+  })
+
+  resp <- req_perform_connection(req_200)
+  expect_equal(resp_stream_lines(resp), "a")
+  expect_equal(resp_stream_lines(resp), "b")
+  close(resp)
+
+  expect_error(req_perform_connection(req_404), class = "httr2_http_404")
+})
+
+
+# StreamingBody --------------------------------------------------------------
+
+test_that("validates its input", {
+  expect_snapshot(error = TRUE, {
+    StreamingBody$new(1)
+  })
 })