+++ NEWS.md
@@ -1,5 +1,6 @@
 # httr2 (development version)
 
+* `req_perform_iterative()`, `req_perform_sequential()`, `req_perform_parallel()`, `req_perform_promise()`, `req_perform_connection()` now support mocking (#651). To mock a streaming connection, create a response with a body that's a `StreamingBody` instance.
 * `req_oauth_device()` gains an `open_browser` argument that lets you take control of whether a browser is opened or the URL is printed (@plietar, #763)
 * `req_perform_parallel()` handles `progress` argument consistently with other functions (#726).
 * `req_error()` errors with long bodies are now correctly wrapped (#727).